Refactor - moved OozeToken to shared area
