+++ Mage.Sets/src/mage/sets/gatecrash/OozeFlux.java
@@ -42,6 +42,7 @@ import mage.cards.CardImpl;
 import mage.counters.CounterType;
 import mage.filter.common.FilterControlledCreaturePermanent;
 import mage.game.Game;
+import mage.game.permanent.token.OozeToken;
 import mage.game.permanent.token.Token;
 
 /**
