+++ Mage.Sets/src/mage/sets/returntoravnica/SlimeMolding.java
@@ -27,17 +27,16 @@
  */
 package mage.sets.returntoravnica;
 
-import java.util.UUID;
-
-import mage.constants.CardType;
-import mage.constants.Rarity;
-import mage.MageInt;
 import mage.abilities.Ability;
 import mage.abilities.effects.OneShotEffect;
 import mage.cards.CardImpl;
+import mage.constants.CardType;
 import mage.constants.Outcome;
+import mage.constants.Rarity;
 import mage.game.Game;
-import mage.game.permanent.token.Token;
+import mage.game.permanent.token.OozeToken;
+
+import java.util.UUID;
 
 /**
  *
@@ -89,18 +88,5 @@ class SlimeMoldingEffect extends OneShotEffect {
     @Override
     public SlimeMoldingEffect copy() {
         return new SlimeMoldingEffect(this);
-    }
-}
-
-class OozeToken extends Token {
-
-    public OozeToken() {
-        super("Ooze", "X/X green ooze creature token");
-
-        cardType.add(CardType.CREATURE);
-        color.setGreen(true);
-        subtype.add("Ooze");
-        power = new MageInt(0);
-        toughness = new MageInt(0);
     }
 }
