+++ Mage.Sets/src/mage/sets/riseoftheeldrazi/GelatinousGenesis.java
@@ -31,13 +31,12 @@ import java.util.UUID;
 
 import mage.constants.CardType;
 import mage.constants.Rarity;
-import mage.MageInt;
 import mage.abilities.Ability;
 import mage.abilities.effects.OneShotEffect;
 import mage.cards.CardImpl;
 import mage.constants.Outcome;
 import mage.game.Game;
-import mage.game.permanent.token.Token;
+import mage.game.permanent.token.OozeToken;
 
 /**
  *
@@ -92,14 +91,3 @@ class GelatinousGenesisEffect extends OneShotEffect {
     }
 }
 
-class OozeToken extends Token {
-
-    public OozeToken() {
-        super("Ooze", "X/X green ooze creature token");
-        cardType.add(CardType.CREATURE);
-        color.setGreen(true);
-        subtype.add("Ooze");
-        power = new MageInt(0);
-        toughness = new MageInt(0);
-    }
-}
